requirements.txt
flask
gunicorn
line-bot-sdk
requests
tinydb